src/log-to-kql/requirements.txt
semantic-kernel
promptflow 
promptflow-tools
python-dotenv
azure-monitor-query
azure.identity